Refactor v1 prep (#28)
* Make versioneer unvendored

* Remove testing module

* Remove serializers module

* Remove filesystems module

* Update project build definitions to be pyproject.toml based

* Move package listing to more universally understood requirements.txt

* Move to ruff formatting

* Clean up docs

* Fix: set ruff formatting config

* Refactor static code analysis action

* Refactor package/docs build and publish

* Refactor unit test action

* Fix conftest to treat harness as context manager

* Remove support for Python 3.13

* Add Python 3.12 to PyPI classifiers

* Don't invoke pytest directly, instead through python module optional argument

* Fix dependency install during test action

* Split package and docs build and fix dependency install
