--- .flake8
@@ -1,6 +1,0 @@
-[flake8]
-ignore = E203, E266, E501, W503, F403, F401,
-exclude = versioneer.py, src/prefecto/_version.py
-max-line-length = 88
-max-complexity = 18
-select = B,C,E,F,W,T4,B9
