+++ .github/workflows/publish-docs.yml
@@ -1,30 +1,36 @@
-name: Publish docs
+name: Build & Publish Package
 
 on:
-  workflow_dispatch
+  release:
+    types: [published]
+  workflow_dispatch:
 
 jobs:
-  build-and-publish-docs:
-    name: Build and publish docs
-    runs-on: ubuntu-latest
-
-    steps:
-      - uses: actions/checkout@v3
-      - name: Set up Python 3.10
-        uses: actions/setup-python@v4
-        with:
-          python-version: "3.10"
-          cache: pip
-          cache-dependency-path: Pipfile
-
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          python -m pip install --upgrade --upgrade-strategy eager -e ".[dev]"
-          mkdocs build
-      
-      - name: Publish docs
-        uses: JamesIves/github-pages-deploy-action@v4.4.2
-        with:
-          branch: docs
-          folder: site
+    build-and-publish-docs:
+        name: Build and publish docs
+        runs-on: ubuntu-latest
+        if: ${{ github.event.inputs.docs == 'true' }}
+    
+        steps:
+          - uses: actions/checkout@v3
+          - name: Set up Python 3.10
+            uses: actions/setup-python@v4
+            with:
+              python-version: "3.10"
+              cache: pip
+              cache-dependency-path: requirements*.txt
+    
+          - name: Install dependencies
+            run: |
+              python -m pip install --upgrade pip
+              python -m pip install -r requirements.txt
+              python -m pip install -r requirements-dev.txt
+    
+          - name: Build docs
+            run: mkdocs build
+    
+          - name: Publish docs
+            uses: JamesIves/github-pages-deploy-action@v4.4.2
+            with:
+              branch: docs
+              folder: site
