+++ .github/workflows/publish-package.yml
@@ -1,25 +1,28 @@
-name: Build & Release
+name: Build & Publish Package
 
 on:
-  push:
-    tags:
-      - "v*"
+  release:
+    types: [published]
 
 jobs:
   build-release:
     name: Build Release
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v4
+
       - name: Set up Python
         uses: actions/setup-python@v4
         with:
           python-version: "3.10"
+          cache: pip
+          cache-dependency-path: requirements*.txt
 
       - name: Install packages
         run: |
           python -m pip install --upgrade pip build
-          python -m pip install --upgrade --upgrade-strategy eager -e ".[dev]"
+          python -m pip install -r requirements.txt
+          python -m pip install -r requirements-dev.txt
 
       - name: Build a binary wheel and a source tarball
         run: |
@@ -45,34 +48,7 @@ jobs:
           path: './dist'
 
       - name: Publish distribution to PyPI
-        uses: pypa/gh-action-pypi-publish@master
+        uses: pypa/gh-action-pypi-publish@release/v1
         with:
           password: ${{ secrets.PYPI_API_TOKEN }}
           verbose: true
-
-  build-and-publish-docs:
-    name: Build and publish docs
-    needs: [build-release, publish-release]
-    runs-on: ubuntu-latest
-
-    steps:
-      - uses: actions/checkout@v3
-
-      - name: Set up Python 3.10
-        uses: actions/setup-python@v4
-        with:
-          python-version: "3.10"
-          cache: pip
-          cache-dependency-path: Pipfile
-
-      - name: Build docs
-        run: |
-          python -m pip install --upgrade pip pyOpenSSL
-          python -m pip install --upgrade --upgrade-strategy eager -e ".[dev]"
-          mkdocs build
-
-      - name: Publish docs
-        uses: JamesIves/github-pages-deploy-action@v4.4.2
-        with:
-          branch: docs
-          folder: site
