+++ .github/workflows/pytest.yml
@@ -3,18 +3,23 @@ name: Tests
 on:
   pull_request:
     paths:
-      - '**.py'
+      - '.github/workflows/pytest.yml'
+      - 'requirements.txt'
+      - 'requirements-dev.txt'
+      - 'src/**'
+      - 'tests/**'
 
 jobs:
-  run-tests:
-    name: Run Tests
+  unit-tests:
+    name: Unit Tests
     runs-on: ubuntu-latest
     strategy:
       matrix:
         python-version:
           - "3.10"
           - "3.11"
-      fail-fast: false
+          - "3.12"
+
     steps:
       - uses: actions/checkout@v3
 
@@ -22,22 +27,15 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
+          cache-dependency-path: requirements*.txt
           cache: pip
-          cache-dependency-path: Pipfile
-
+  
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          python -m pip install --upgrade --upgrade-strategy eager -e ".[dev,pandas,polars,moto]"
+          python -m pip install -r requirements.txt
+          python -m pip install -r requirements-dev.txt
 
       - name: Run tests
-        env:
-          PREFECT_SERVER_DATABASE_CONNECTION_URL: "sqlite+aiosqlite:///./collection-tests.db"
         run: |
-          prefect server database reset -y
-          coverage run --branch -m pytest tests -vv
-          coverage report
-
-      - name: Run mkdocs build
-        run: |
-          mkdocs build --verbose --clean
+          python -m pytest tests/ -vv
