+++ .github/workflows/static-analysis.yml
@@ -3,23 +3,15 @@ name: Static analysis
 on:
   pull_request:
     paths:
-      - '**.py'
+      - .github/workflows/static-analysis.yml
+      - .pre-commit-config.yaml
+      - .ruff.toml
+      - pyproject.toml
+      - src/**
+      - tests/**
 
 permissions:
   contents: read
-
-# Limit concurrency by workflow/branch combination.
-#
-# For pull request builds, pushing additional changes to the
-# branch will cancel prior in-progress and pending builds.
-#
-# For builds triggered on a branch push, additional changes
-# will wait for prior builds to complete before starting.
-#
-# https://docs.github.com/en/actions/using-jobs/using-concurrency
-concurrency:
-  group: ${{ github.workflow }}-${{ github.ref }}
-  cancel-in-progress: ${{ github.event_name == 'pull_request' }}
 
 jobs:
   pre-commit-checks:
