+++ .gitignore
@@ -1,6 +1,3 @@
-# Local package management
-Pipfile.lock
-
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
