+++ .pre-commit-config.yaml
@@ -1,22 +1,17 @@
 repos:
-  - repo: https://github.com/pycqa/isort
-    rev: 5.12.0
+  - repo: https://github.com/charliermarsh/ruff-pre-commit
+    rev: "v0.3.7"
     hooks:
-      - id: isort
-        args: ["--profile", "black","--filter-files"]
+      - id: ruff
         language_version: python3
+        args: [--fix, --exit-non-zero-on-fix, --show-fixes]
   - repo: https://github.com/psf/black
-    rev: 22.3.0
+    rev: 24.4.0
     hooks:
       - id: black
-        exclude: src/prefecto/_version.py
         language_version: python3
-  - repo: https://github.com/pycqa/flake8
-    rev: 4.0.1
-    hooks:
-      - id: flake8
   - repo: https://github.com/econchick/interrogate
-    rev: 1.5.0
+    rev: 1.7.0
     hooks:
       - id: interrogate
         args: [-vv]
