--- Pipfile
@@ -1,39 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-prefect = ">=2.10"
-
-[dev-packages]
-pre-commit = ">=3.3.3"
-flake8 = ">=6.0.0"
-black = ">=23.3.0"
-isort = ">=5.12.0"
-pytest = ">=7.3.2"
-interrogate = ">=1.5.0"
-mkdocs = ">=1.4.3"
-versioneer = ">=0.28.0"
-coverage = ">=7.2.7"
-mkdocstrings = {version = ">=0.22.0", extras = ["python"]}
-
-[requires]
-python_version = "3.11"
-
-[requests]
-requests = "*"
-
-[pandas]
-pandas = "*"
-pyarrow = "*"
-openpyxl = "*"
-
-[polars]
-polars = "*"
-xlsxwriter = "*"
-xlsx2csv = "*"
-
-[moto]
-moto = {extras = ["s3"], version = "*"}
-boto3 = "*"
