--- docs/filesystems.md
@@ -1,11 +1,0 @@
-# Filesystems
-
-Tools to extend Prefect file system behavior.
-
-## create_child
-
-::: src.prefecto.filesystems.create_child
-
-## task_persistence_subfolder
-
-::: src.prefecto.filesystems.task_persistence_subfolder
