--- docs/serializers/serialization.md
@@ -1,28 +1,0 @@
-# Serialization
-
-Library for serializing various data types.
-
-## ExtendedSerializer
-
-Type: `ext`
-
-::: src.prefecto.serializers.core.ExtendedSerializer
-    options:
-        members:
-            - get_method
-            - dumps
-            - loads
-
-## Method
-
-Base class for creating serialization methods.
-
-::: src.prefecto.serializers.core.Method
-    options:
-        members:
-            - read
-            - write
-
-## get_method
-
-::: src.prefecto.serializers.core.get_method
