--- docs/testing/s3.md
@@ -1,5 +1,0 @@
-# S3
-
-## mock_bucket
-
-::: src.prefecto.testing.s3.mock_bucket
