+++ requirements-dev.txt
@@ -1,0 +1,5 @@
+mkdocs>=1.4.3
+mkdocstrings-python>=1.1.2
+pre-commit>=3.5.0
+pytest>=7.4.3
+versioneer>=0.29
