--- setup.cfg
@@ -1,14 +1,0 @@
-[options]
-package_dir=
-    =src
-zip_safe = True
-include_package_data = True
-
-[versioneer]
-VCS = git
-style = pep440
-versionfile_source = src/prefecto/_version.py
-versionfile_build = prefecto/_version.py
-tag_prefix = v
-parentdir_prefix =
-
