+++ src/prefecto/logging.py
@@ -1,6 +1,7 @@
 """
 Prefect logging utilities.
 """
+
 from __future__ import annotations
 
 import logging
