+++ src/prefecto/states.py
@@ -2,6 +2,7 @@
 Tools to improve Prefect states.
 
 """
+
 from __future__ import annotations
 
 from prefect import states
