+++ tests/conftest.py
@@ -1,6 +1,7 @@
 """
 Unit test configuration file.
 """
+
 from __future__ import annotations
 
 import pytest
@@ -10,7 +11,5 @@ from prefect.testing.utilities import prefect_test_harness
 @pytest.fixture
 def harness():
     """Return a `prefect_test_harness`."""
-    try:
-        prefect_test_harness()
-    finally:
-        pass
+    with prefect_test_harness():
+        yield
