+++ tests/test_logging.py
@@ -1,6 +1,7 @@
 """
 Tests for the logging module.
 """
+
 from __future__ import annotations
 
 import logging
